# Conflicts:
#	src/consts/const_global.js
#	src/node/sockets/node-clients/service/discovery/fallbacks/fallback_nodes_list.js

+++ src/common/blockchain/ppow-blockchain/blockchain/prover/proofs/PPoW-Blockchain-Proof-Basic.js
@@ -4,6 +4,19 @@ import consts from 'consts/const_global'
 
 class PPoWBlockchainProofBasic{
 
+    constructor(blockchain, blocks){
+
+        this.date = new Date();
+
+        this.blockchain = blockchain;
+
+        this.blocks = blocks;
+        this.blocksIndex = {};
+
+        this.hash = undefined;
+
+    }
+
     destroyProof(){
 
         if (!this.blockchain.agent.light) {
@@ -27,7 +40,7 @@ class PPoWBlockchainProofBasic{
             if (!found) {
 
                 // avoid destroying real blocks
-                if (typeof this.blocks[i].destroyBlock === "function")
+                if (this.blocks[i] !== undefined && typeof this.blocks[i].destroyBlock === "function")
                     this.blocks[i].destroyBlock();
 
             }
@@ -40,16 +53,7 @@ class PPoWBlockchainProofBasic{
 
     }
 
-    constructor(blockchain, blocks){
 
-        this.date = new Date();
-
-        this.blockchain = blockchain;
-        this.blocks = blocks;
-
-        this.hash = undefined;
-
-    }
 
     getProofHeaders(starting, length){
 
@@ -144,6 +148,17 @@ class PPoWBlockchainProofBasic{
             }
 
         return null;
+    }
+
+
+    push(block){
+
+        if (this.blocksIndex[block.height] === block)
+            return;
+
+        this.blocks.push(block);
+        this.blocksIndex[block.height] = block;
+
     }
 
 }
